RL/RoadLogica-HKJC/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.3.0
billiard==3.3.0.23
celery==3.1.25
certifi==2019.11.28
chardet==3.0.4
cloud-init==19.1.4
configobj==5.0.6
cryptography==2.6.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.1.15
django-import-export==2.3.0
entrypoints==0.3
et-xmlfile==1.0.1
furl==2.1.0
idna==2.8
importlib-metadata==1.6.1
jdcal==1.4.1
Jinja2==2.11.2
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
keyring==17.1.1
keyrings.alt==3.1.1
kombu==5.0.2
lxml==4.5.0
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.4
orderedmultidict==1.0.1
psycopg2-binary==2.8.5
pycrypto==2.6.1
PyGObject==3.30.4
pyrsistent==0.16.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
redis==2.10.6
requests==2.22.0
SecretStorage==2.3.1
selenium==3.141.0
simpleui==3.9
six==1.14.0
tablib==2.0.0
urllib3==1.25.8
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
